refactor: Phase 3: Cleanup
